+++ packages/design-system/src/components/Page/PageContent.tsx
@@ -19,7 +19,7 @@ export const PageContent: FC<PageContentProps> = ({ ref, children, className, ..
       ref={ref}
       data-testid={testId}
       data-slot='page-content'
-      className={cn('flex-1 min-h-0 overflow-auto px-16', className)}
+      className={cn('flex-1 min-h-0 overflow-auto px-16 pt-8 pb-16', className)}
     >
       {children}
     </div>